requirements.txt
SoundFile
torchaudio
munch
torch
pydub
pyyaml
librosa
nltk
matplotlib
accelerate
transformers
einops
einops-exts
tqdm
typing
typing-extensions
git+https://github.com/resemble-ai/monotonic_align.git # or resemble-monotonic-align
gradio
phonemizer
cached-path
tortoise-tts # for the Gradio demo, splitting text
flask # for api
markdown # for api
flask-cors
eventlet